app/requirements.txt
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
docker==4.3.1
english==2020.7.0
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
neotime==1.7.4
packaging==20.4
pansi==2020.7.3
prompt-toolkit==2.0.10
py2neo==2020.0.0
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
urllib3==1.25.11
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
